==> MODULO 3 - copia/SESION 8 - 14 JULIO/1 conectandose-sqlite3.ipynb <==
## Ejemplo 1: Conectándose a una base de datos

### 1. Objetivos:
    - Usar un conector de python para conectarse a una base de datos.
 
### 2. Desarrollo:

#### a) Realizar una conexión y obtener un índice o cursor

Para conectarnos necesitamos primero importar el módulo o paquete para la base de datos en particular, en este caso SQLite3 y el módulo se llama `sqlite3`, así que importamos el módulo:

import sqlite3

Para una conexión a SQLite necesitamos saber el nombre del archivo de la base de datos, algo como `mi_datos.sqlite3`.

Para una conexión MySQL, Postgres, Oracle u otras necesitamos conocer las siguiente información:

1. Host (localhost, 127.0.0.1, servidor.dominio.com)
2. Puerto (3306 u otro)
3. Usuario (fulanito)
4. Contraseña (tiamida)
5. Nombre de la base de datos (camino_al_exito)

Todos esos datos los usamos para crear una conexión usando para ello la siguiente función del módulo `sqlite3`:

`sqlite3.connect(-nombre de la base de datos-)`

entonces vamos a realizar una conexión a la base de datos SQLite3 llamada `movielens.sqlite3`:

conn = sqlite3.connect("movielens.sqlite3")

conn

Ahora, para poder realizar modificaciones o consultas a nuestra base de datos, tenemos que crear un objeto llamado índice o `cursor` usando:

`mi_conexion.cursor()`

cur = conn.cursor()

cur

Éste objeto lo podemos usar para realizar cualquier consulta en lenguaje SQL. Por ejemplo, veamos qué tablas existen en nuestra base de datos ejecutando consultas SQL usando:

`mi_cursor.execute(-sql-)`

Nota: En el caso de SQLite3, para obtener la lista de tablas se puede hacer uso de la instrucción SQL:

`SELECT name FROM sqlite_master WHERE type='table';`

en el caso de MariaDB o MySQL se podría usar:

`SHOW TABLES;`

entonces ejecutemos la instrucción:

cur.execute("SELECT name FROM sqlite_master WHERE type='table'")

Después de ejecutar el comando, tenemos que extraer la información con la función:

`mi_cursor.fetch()` o
`mi_cursor.fetchall()`

resultados = cur.fetchall()

resultados

Listo, ahora sabemos que tenemos 5 tablas en nuestra base de datos!

Por último, es importante cerrar nuestro `cursor` y conexión a la base de datos para no utilizar memoria extra en nuestra computadora:

`mi_cursor.close()` y
`mi_conexion.close()`

cur.close()

conn.close()
