==> MODULO 3 - copia/SESION 8 - 14 JULIO/2 tablas_a_dataframes-sqlite3.ipynb <==
## Ejemplo 2: Consultas a la base de datos y construcción de `DataFrames`

### 1. Objetivos:
    - Realizar consultas a la base de datos y construir un `DataFrame` para cada tabla.
 
---
    
### 2. Desarrollo:

#### a) Tablas a `DataFrames`

import pandas as pd
import sqlite3

Realizamos la conexión y obtenemos nuestro índice (cursor) para poder realizar operaciones con la base de datos:

bd = "movielens.sqlite3"
conn = sqlite3.connect(bd)
cur = conn.cursor()

cur

Ahora vamos a obtener la lista de todos los registros de la tabla `users` usando la función `execute(-sql-)` y la instrucción SQL:

`SELECT * FROM users`

sql = "SELECT * FROM users"
cur.execute(sql)
cur

Ahora hay que obtener todos los resultados de la consulta con:
    
`mi_cursor.fetchall()`

resultados = cur.fetchall()

resultados [:5]

Si recuerdas el último ejemplo, `fetchall` nos regresa una `lista` de `tuplas`. Cada `tupla` representa una fila de nuestro conjunto de datos, por ejemplo obtén el primer registro se podría obtener como:

resultados [0]

O podríamos obtener el género del primer usuario con:

resultados[0][1]

Algo muy util es obtener la lista de los nombres de las columnas de la tabla consultada, para eso se hacer uso del atributo:

`mi_cursor.description`

cur.description

y aplicando listas de compresión se puede obtener la lista deseada:

columnas = [t[0] for t in cur.description]  # un ejemplo t= ('user_id', None, None, None, None, None, None)

columnas

Afortunadamente `pandas` puede recibir justamente `listas` de `tuplas` como ingredientes para construir `DataFrames`. Sólo hace falta indicarle el nombre de las columnas. Los nombres de las columnas están especificados en el [archivo Readme.md](../../Datasets/MovieLens/Readme.md) que venía incluido con el dataset:

Así que ahora se crea el dataframe usando la forma:

`pd.DataFrame(-lista de registros-,
    columns=-lista de nombres de columnas-)`

df = pd.DataFrame(resultados, columns=columnas)

df.head()

Otra manera de realizar esta operación de una manera más compacta es utilizando la siguiente función de Pandas:

`pd.read_sql(-consulta sql-, -variable de conexión a la BD-)`

así que vamos a obtener justo el mismo dataframe y lu guardaremos en la viable `df_1`:

df_1 = pd.read_sql(sql, conn)

df_1.head()

Sería una buena idea convertir la columna `user_id` en índice para no tener información redundante usando:

`dataframe.set_index(-columna-, drop=True)`

y el resultado lo guardamos en la variable `df_2`:

df_2 = df.set_index("user_id", drop=True)

df_2.head(3)

¡Listo! Vamos a guardar nuestros `DataFrames` en el archivo `users.csv` para que no haga falta volver a extraerlos de la base de datos. Ya podemos decirle adiós a nuestro fiel MySQL. ¡Chao, bellisimo!

df_2.to_csv("users.csv")

Ahora hagamos lo mismo para la tabla `occupations`, obtengamos todos los registros, examinemos, obtengamos los nombres de las columnas, creemos el dataframe, reasignemos índices en caso de ser necesario y guardemos el resultado en el archivo `occupations.csv`

ocupation = pd.read_sql("SELECT * FROM occupations", conn)
ocupation

Haaa y no olvides cerrar tu índice y tu conexión:

occupations = ocupation.set_index("occupation_id", drop=True)
occupations

occupations.to_csv("occupations.csv")

---
---

## Reto 2: Convertir tablas en MySQL a `DataFrames` de `pandas`

### 1. Objetivos:
    - Solicitar todos los datos de las tablas que están almacenadas en nuestra base de datos, convertirlos a `DataFrames` y guardarlos.
    
---
    
### 2. Desarrollo:

#### a) Tablas a `DataFrames`

Ya que tenemos nuestra conexión funcionando adecuadamente, vamos a utilizarla para realizar consultas a las base de datos y construir una base de datos local. Tu Reto consiste en los siguientes pasos:

1. Vuelve a establecer la conexión a la base de datos
2. Usando el comando `SELECT * FROM nombre_de_tabla`, realiza consultas a cada una de las 5 tablas que existen en la base de datos.
3. Crea un `DataFrame` por cada tabla que obtuviste. Para asignarle los nombres de las columnas correctamente, revisa el archivo [Readme.md](../../Datasets/MovieLens/Readme.md) donde está contenida toda esa información.
4. Asegúrate de que el índice sea adecuado en cada `DataFrame`. En los casos en los que haya datos redundantes, convierte una de las columnas en índice.
5. Si lo deseas ordena las columnas de la manera en la que mejor te parezca.
6. Guarda tus `DataFrames` en formato .csv para utilizarlos en los siguientes Retos.

> **NOTA IMPORTANTE**: La tabla movies es un poco complicada porque contiene muchos signos distintos. Tanto en la columna de nombre de película como la de género, encontramos signos como `,`, `:`, `.`, `|`. Esto hace un poco complicado el almacenamiento y lectura de este archivo. Si elijes guardar este archivo como un .csv separado por comas (`,`), a la hora de leerlo de regreso, `pandas` puede confundirse y pensar que el título de una película que contiene comas constituye dos columnas. Por esta razón, te recomiendo que la tabla `movies` la guardes agregando un separador poco convencional como `sep='$'`. De esta manera será muchísimo más fácil leer tu archivo de regreso usando ese separador.

...

Compara con tus compañeros y revisa con la experta para que todos estén seguros de que tienen sus `DataFrames` estructurados de la manera correcta y que sus archivos .csv fueron creados exitosamente. Vamos a utilizar estos archivos en los Retos siguientes, así que es muy importante que tus datos estén estructurados adecuadamente.
